[TASK] Implemented a ComparisonService and a base route for Insert Logic
