+++ Configuration/Backend/AjaxRoutes.php
@@ -11,5 +11,10 @@ return [
     'cfcookiemanager_ajax_updatedataset' => [
         'path' => '/cf-cookiemanager/ajax/update-dataset',
         'target' => BackendAjax\UpdateCheckController::class . '::updateDatasetAction',
+    ],
+
+    'cfcookiemanager_ajax_insertdataset' => [
+        'path' => '/cf-cookiemanager/ajax/insert-dataset',
+        'target' => BackendAjax\UpdateCheckController::class . '::insertDatasetAction',
     ]
 ];
