+++ Resources/Public/Backend/Css/CookieSettings.css
@@ -488,4 +488,9 @@ table.dataTable {
   color: white;
 }
 
+.cf-cookiemanager-insert {
+  background-color: #28a745;
+  color: white;
+}
+
 /*# sourceMappingURL=CookieSettings.css.map */
