+++ Resources/Public/Backend/Css/CookieSettings.scss
@@ -584,3 +584,9 @@ table.dataTable {
   color: white;
 }
 
+.cf-cookiemanager-insert {
+  background-color: #28a745;
+  color: white;
+}
+
+
